#624: implemented requested changes moved testCheckPluginInstallation from IdeToolCommandletTest to IntellijTest moved testCheckEditionConflictInstallation from IdeToolCommandletTest to IntellijTest Adjusted IntellijTest testIntellijPluginInstallWithCustomRepoUrl to use install instead of run
